# [ez][HUD] Move joins around to make queries faster