Move some join orderings around and use lookup joins

Is this necessary? No, but I noticed that the avg time to run queries decreased a lot after https://github.com/pytorch/test-infra/pull/5026, even for queries that were not the ones changed in the PR.  I think reducing the time/compute of one query makes it faster to run other queries at the same time

In the query editor console:
master_commit_red_avg was about 2500ms -> 400ms
master_commit_red was about 1200ms -> 300ms